+++ knowledge/example_code/FixtureAdapter.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace AKlump\LiveDevPorter\Processors;
+
+use AKlump\LiveDevPorter\Fixture\Permissions as PermissionsFixture;
+use AKlump\LiveDevPorter\FixtureFramework\AbstractFixtureAdapter;
+use AKlump\LiveDevPorter\Traits\LocaldevFixtureTrait;
+
+final class Permissions extends AbstractFixtureAdapter {
+
+  use LocaldevFixtureTrait;
+
+  protected function getFixtureClass(): string {
+    return PermissionsFixture::class;
+  }
+
+}